mysql: docs + site + audit follow-ups (§2.10 wrap-up)
Documentation, site, and two small fixes from a C/JS/Lua audit of the
MySQL backend and stdlib port.

CLAUDE.md: document the MySQL/MariaDB backend. HL_ENABLE_MYSQL flag row +
the three-flag HL_ENABLE_DB umbrella; the three-backend DSN-scheme selector;
a "MySQL/MariaDB specifics" subsection (auth: native + caching_sha2 with the
TLS full-auth path and the deferred ed25519; TLS/sslmode; text vs binary
prepared queries + trailing-nil param padding; dialect INSERT IGNORE / ON
DUPLICATE KEY / information_schema / backtick quoting / the CREATE INDEX IF
NOT EXISTS shim; portable VARCHAR-key stdlib DDL; SQLite-only udf/search);
updated flag combinations + the test-suite table (test_mysqlwire 19,
test_mysql_conn 10).

site/index.html: surface the multi-backend DB support (the site described
storage as SQLite-only). The hero and the "Compute & data" card now note
SQLite plus the pure-C PostgreSQL and MySQL/MariaDB wire clients. The
reproducible-builds and minisign-signed-installer messaging was already
accurate and comprehensive (trust-chain Tier 4 + the install/verify block),
so it was left as-is.

Audit fixes (no Critical/High findings across the three audits):
- cap/mysql_conn.c: binary TIME decode computed `(int)(days*24)+h` where days
  is a server-controlled u32; a hostile value could signed-overflow int
  (UBSan-visible). Now computed in int64_t. ASan/UBSan-clean.
- cap/db_mysql.c: documented that the CREATE INDEX IF NOT EXISTS shim's loose
  substring match is safe because its only callers are trusted stdlib index
  DDL (never app SQL), so it is not an injection surface.

The audits also flagged (as latent, not regressions) that the client-supplied
idempotency key and inbox message_id are now VARCHAR(255) and could truncate
on MySQL; addressing that is a separate follow-up (it needs an input-length
policy that doesn't regress existing SQLite/Postgres apps).

+++ src/hull/cap/mysql_conn.c
@@ -804,15 +804,17 @@ static void decode_binary_value(HlMyCursor *rc, uint8_t type, HlMyVal *out)
         if (len >= 8)  { neg = hl_my_get_u8(rc); days = hl_my_get_u32(rc);
                          h = hl_my_get_u8(rc); mi = hl_my_get_u8(rc); se = hl_my_get_u8(rc); }
         if (len >= 12) { us = hl_my_get_u32(rc); }
-        int hours = (int)(days * 24) + h;
+        /* days is server-controlled (u32); compute in int64 so a hostile value
+         * cannot signed-overflow int (UBSan-visible on untrusted input). */
+        int64_t hours = (int64_t)days * 24 + h;
         if (len == 0)
             snprintf(out->tbuf, sizeof out->tbuf, "00:00:00");
         else if (us)
-            snprintf(out->tbuf, sizeof out->tbuf, "%s%02d:%02d:%02d.%06u",
-                     neg ? "-" : "", hours, mi, se, us);
+            snprintf(out->tbuf, sizeof out->tbuf, "%s%02lld:%02d:%02d.%06u",
+                     neg ? "-" : "", (long long)hours, mi, se, us);
         else
-            snprintf(out->tbuf, sizeof out->tbuf, "%s%02d:%02d:%02d",
-                     neg ? "-" : "", hours, mi, se);
+            snprintf(out->tbuf, sizeof out->tbuf, "%s%02lld:%02d:%02d",
+                     neg ? "-" : "", (long long)hours, mi, se);
         out->kind = HL_MY_VAL_STR;
         out->v.s.ptr = out->tbuf; out->v.s.len = strlen(out->tbuf);
         return;
